+++ src/components/sections/elements/Dates.vue
@@ -53,10 +53,11 @@ watch(
   <SectionBorder :type="SectionType.Element">
     <SectionTitle
       :type="SectionType.Element"
-      version="2.2"
+      :stability="Stability.Stable"
+      version="2.3"
       anchor="dates"
       title="Dates"
-      :stability="Stability.Experimental"
+      guidance-href="https://gitlab.data.bas.ac.uk/MAGIC/mapping-coordination/-/wikis/metadata-completion-guidance#dates"
     />
     <TwoColumn>
       <template #left>