+++ src/components/sections/elements/Identifiers.vue
@@ -71,10 +71,11 @@ watch(
   <SectionBorder :type="SectionType.Element">
     <SectionTitle
       :type="SectionType.Element"
-      version="6.0"
-      :stability="Stability.Experimental"
+      :stability="Stability.Stable"
+      version="6.1"
       anchor="identifiers"
       title="Identifiers"
+      guidance-href="https://gitlab.data.bas.ac.uk/MAGIC/mapping-coordination/-/wikis/metadata-completion-guidance#identifiers"
       :depends-on="dependantSections"
     />
     <TwoColumn>