+++ src/components/sections/elements/ResourceType.vue
@@ -34,10 +34,11 @@ watch(
   <SectionBorder :type="SectionType.Element">
     <SectionTitle
       :type="SectionType.Element"
-      version="2.3"
       :stability="Stability.Stable"
+      version="2.4"
       anchor="resource-type"
       title="Resource Type"
+      guidance-href="https://gitlab.data.bas.ac.uk/MAGIC/mapping-coordination/-/wikis/metadata-completion-guidance#resource-type"
     />
     <TwoColumn>
       <template #left>