+++ src/components/sections/elements/Thumbnails.vue
@@ -57,10 +57,11 @@ watch(
   <SectionBorder :type="SectionType.Element">
     <SectionTitle
       :type="SectionType.Element"
-      version="1.0"
-      :stability="Stability.Experimental"
+      :stability="Stability.Stable"
+      version="1.1"
       anchor="thumbnails"
       title="Thumbnails"
+      guidance-href="https://gitlab.data.bas.ac.uk/MAGIC/mapping-coordination/-/wikis/metadata-completion-guidance#thumbnails"
       :data-file-href="['formats.json', 'thumbnails.json']"
       :depends-on="dependantSections"
     />